NormalizeTF V1.1
Updated version outputs unique list of TFs (no duplicates). We are now
also using a cutoff of TFs present in 2/5 lists in order to avoid
exclusion of TFs not in old databases (TFCat, DBD) or that were not
among the small selection of TFs tested by ENCODE. See script for
details.