VALIDATE
In this version, I have updated the output of VALIDATE module. I
included an example of generating a .csv file
contains the validation results and it can be use to generate VALIDATE
when we get result from RANK module. Please see VALIDATE page on the
student wiki for more information